requirements.txt
autopep8==1.5.3
braintree==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
Django==2.2
django-admin-thumbnails==0.2.5
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
idna==2.9
Pillow==7.1.2
pycodestyle==2.6.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
stripe==2.48.0
toml==0.10.1
urllib3==1.25.9
